requirements.txt
imageio==2.31.6
matplotlib==3.8.1
numpy==1.23.5
opencv-python==4.8.1.78
pandas==2.1.3
Pillow==9.5.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
timm==0.3.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.66.1
